Beauty script file and setup auto update mac

+++ main.py
@@ -6,6 +6,50 @@ import subprocess
 url = "http://158.160.166.58:30002/messages"
 headers = "Content-Type: application/json"
 
+def copy_config(dir):
+    with open(dir+"/QtCreator.ini", "a", encoding="utf-8") as dest:
+            dest.write("\n"+code_to_add)
+    shutil.copy(dir+"/QtCreator.ini", dir+"/QtCreatorBackup.ini")
+    source_folder = "beautifier"
+    destination_folder = dir+"/qtcreator"
+    shutil.copytree(source_folder, os.path.join(destination_folder, os.path.basename(source_folder)), dirs_exist_ok=True)
+    source_folder = "clang-format"
+    shutil.copytree(source_folder, os.path.join(destination_folder, os.path.basename(source_folder)), dirs_exist_ok=True)
+
+
+def install_updates(dir):
+    subprocess.run(['git', '-C', '.', 'reset', '--hard'])
+    subprocess.run(['git', '-C', '.', 'pull'])
+    with open("forConfigUbuntu.ini", "r", encoding="utf-8") as src:
+        code_to_add = src.read()
+    remove_old_config(dir+"/QtCreator.ini", "[Beautifier]", "ForceEnabled=Beautifier")
+    with open(dir+"/QtCreator.ini", "a", encoding="utf-8") as dest:
+        dest.write("\n"+code_to_add)   
+
+def check_install(surname):
+    print("\nДля проверки установки проделайте шаги указанные в файле check.txt, если все успешно, нажмите Y, иначе N")    
+    while(True):
+        user_input = input().strip().upper()
+        if user_input == 'Y':
+            subprocess.run([
+                "curl",
+                "-X", "POST",
+                "-d", f'{{"message": "{surname+":OK"}"}}',
+                "-H", headers,
+                url
+            ])
+            break
+        elif user_input == 'N':
+            subprocess.run([
+                "curl",
+                "-X", "POST",
+                "-d", f'{{"message": "{surname+":FAIL"}"}}',
+                "-H", headers,
+                url
+            ])
+            break
+        else:
+            print("Неверный ввод! Пожалуйста, нажмите Y или N.")
 
 def remove_old_config(file_path, start_marker, end_marker):
     with open(file_path, 'r') as file:
@@ -49,63 +93,35 @@ if platform.system() == "Linux":
         print("Поиск обновлений")
         if check_git_updates():
             print("Найдено обновление")
-            subprocess.run(['git', '-C', '.', 'reset', '--hard'])
-            subprocess.run(['git', '-C', '.', 'pull'])
-            with open("forConfigUbuntu.ini", "r", encoding="utf-8") as src:
-                code_to_add = src.read()
-            remove_old_config(dir+"/QtCreator.ini", "[Beautifier]", "ForceEnabled=Beautifier")
-            with open(dir+"/QtCreator.ini", "a", encoding="utf-8") as dest:
-                dest.write("\n"+code_to_add)
+            install_updates(dir)
             print("Обновление установлено")
         else:
             print("Обновление не найдено")
     else:    
         print("Введите фамилию: ")
         surname = input()
-        
         with open("forConfigUbuntu.ini", "r", encoding="utf-8") as src:
             code_to_add = src.read()
-        with open(dir+"/QtCreator.ini", "a", encoding="utf-8") as dest:
-            dest.write("\n"+code_to_add)
-        shutil.copy(dir+"/QtCreator.ini", dir+"/QtCreatorBackup.ini")
-        source_folder = "beautifier"
-        destination_folder = dir+"/qtcreator"
-        shutil.copytree(source_folder, os.path.join(destination_folder, os.path.basename(source_folder)), dirs_exist_ok=True)
-        source_folder = "clang-format"
-        shutil.copytree(source_folder, os.path.join(destination_folder, os.path.basename(source_folder)), dirs_exist_ok=True)
+        copy_config(dir)
         subprocess.run("sudo apt update && sudo apt upgrade -y && sudo apt install -y curl && sudo apt install -y clazy && sudo apt install -y clang-format && sudo apt install -y cmake && sudo apt-get install -f", shell=True)
-        print("\nДля проверки установки проделайте шаги указанные в файле check.txt, если все успешно, нажмите Y, иначе N")
-        while True:
-            user_input = input().strip().upper()
-            if user_input == 'Y':
-                subprocess.run([
-                    "curl",
-                    "-X", "POST",
-                    "-d", f'{{"message": "{surname+":OK"}"}}',
-                    "-H", headers,
-                    url
-                ])
-                break
-            elif user_input == 'N':
-                subprocess.run([
-                    "curl",
-                    "-X", "POST",
-                    "-d", f'{{"message": "{surname+":FAIL"}"}}',
-                    "-H", headers,
-                    url
-                ])
-                break
-            else:
-                print("Неверный ввод! Пожалуйста, нажмите Y или N.")
+        check_install(surname)
+
 elif platform.system() == "Darwin":
-    subprocess.run("brew update && brew upgrade && brew install llvm && brew install clazy && brew install cmake", shell = True)
     dir = os.getenv('HOME') + "/.config/QtProject"
-    with open("forConfigMac.ini", "r", encoding="utf-8") as src:
-        code_to_add = src.read()
-    with open(dir+"/QtCreator.ini", "a", encoding="utf-8") as dest:
-        dest.write("\n"+code_to_add)
-    source_folder = "beautifier"
-    destination_folder = dir+"/qtcreator"
-    shutil.copytree(source_folder, os.path.join(destination_folder, os.path.basename(source_folder)), dirs_exist_ok=True)
-    source_folder = "clang-format"
-    shutil.copytree(source_folder, os.path.join(destination_folder, os.path.basename(source_folder)), dirs_exist_ok=True)
+    if os.path.exists(dir+"/QtCreatorBackup.ini"):
+        print("Поиск обновлений")
+        if check_git_updates():
+            print("Найдено обновление")
+            install_updates(dir)
+            print("Обновление установлено")
+        else:
+            print("Обновление не найдено")
+    else:    
+        print("Введите фамилию: ")
+        surname = input()
+        with open("forConfigMac.ini", "r", encoding="utf-8") as src:
+            code_to_add = src.read()
+        copy_config(dir)
+        subprocess.run("brew update && brew upgrade && brew install llvm && brew install clazy && brew install cmake", shell = True)
+        check_install(surname)
+
